LayoutTests/storage/indexeddb/factory-deletedatabase-interactions-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test the deleteDatabase call and its interaction with open/setVersion

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "factory-deletedatabase-interactions.html"

TEST: deleteDatabase blocked on open handles
self.testname = 'test1'; self.ver = 1
blockedEventFired = false
indexedDB.open(dbname + testname)

openOnSuccess():
h = event.target.result
indexedDB.deleteDatabase(dbname + testname)

onVersionChange():
old = 1
new = ""
scheduling timeout to close

deleteDatabaseOnBlocked():
blockedEventFired = true

timeoutCallback():
h.close()

deleteDatabaseOnSuccess():
PASS blockedEventFired is true

TEST: deleteDatabase not blocked when handles close immediately
self.testname = 'test2'; self.ver = 1
blockedEventFired = false
indexedDB.open(dbname + testname)

openOnSuccess():
h = event.target.result
indexedDB.deleteDatabase(dbname + testname)

onVersionChange():
old = 1
new = ""
h.close()

deleteDatabaseOnBlocked():
blockedEventFired = true

deleteDatabaseOnSuccess():
FIXME: blocked event should not fire since connection closed. http://webkit.org/b/71130
FAIL blockedEventFired should be false. Was true.

TEST: deleteDatabase is delayed if a VERSION_CHANGE transaction is running
self.testname = 'test3'; self.ver = 1
versionChangeComplete = false
indexedDB.open(dbname + testname)

openOnSuccess():
h = event.target.result
h.setVersion(String(ver++))
indexedDB.deleteDatabase(dbname + testname)

setVersionOnSuccess():

transactionOnComplete():
versionChangeComplete = true

onVersionChange():
old = "1"
new = ""
scheduling timeout to close

deleteDatabaseOnBlocked():

timeoutCallback():
h.close()

deleteDatabaseOnSuccess():
PASS versionChangeComplete is true

TEST: Correct order when there are pending setVersion, delete and open calls.
self.testname = 'test4'; self.ver = 1
setVersionBlockedEventFired = false
versionChangeComplete = false
deleteDatabaseBlockedEventFired = false
deleteDatabaseComplete = false
indexedDB.open(dbname + testname)

openOnSuccess():
h = event.target.result
indexedDB.open(dbname + testname)

h2OpenOnSuccess():
h.setVersion(String(ver++))
indexedDB.deleteDatabase(dbname + testname)

h2OnVersionChange():
old = 1
new = "1"

setVersionOnBlocked():
setVersionBlockedEventFired = true
indexedDB.open(dbname + testname)
h2.close()

setVersionOnSuccess():

transactionOnComplete():
versionChangeComplete = true

onVersionChange():
old = "1"
new = ""
scheduling timeout to close

deleteDatabaseOnBlocked():
deleteDatabaseBlockedEventFired = true

timeoutCallback():
h.close()

deleteDatabaseOnSuccess():
deleteDatabaseComplete = true

h3OpenOnSuccess():
PASS versionChangeComplete is true
PASS blockedEventFired is true
PASS deleteDatabaseComplete is true
PASS successfullyParsed is true

TEST COMPLETE
